slsDetectorUsers reloaded
git-svn-id: file:///afs/psi.ch/project/sls_det_software/svn/slsDetectorSoftware@275 951219d9-93cf-4727-9268-0efd64621fa3

+++ slsDetectorSoftware/slsDetector/slsDetectorUsers.h
@@ -362,7 +362,7 @@ class slsDetectorUsers
      \param userCallback function for postprocessing and saving the data  
   */
   
-   void registerRawDataCallback(int( *userCallback)(float*, void*), void *pArg);
+   void registerRawDataCallback(int( *userCallback)(double*, void*), void *pArg);
   
   /**
      @short register calbback for accessing detector final data
